requirements.txt
httpx==0.26.0
pandas==2.1.4
SQLAlchemy==2.0.25
python-dotenv==1.0.0
psycopg2==2.9.9